Exercises 4.17-4.21 Token authorization

+++ part4/bloglist/controllers/blogs.js
@@ -1,8 +1,11 @@
 const blogsRouter = require("express").Router();
+const jwt = require("jsonwebtoken");
 const Blog = require("../models/blog");
+const User = require("../models/user");
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog
+    .find({}).populate("author");
   response.json(blogs);
 });
   
@@ -13,13 +16,22 @@ blogsRouter.post("/", async (request, response, next) => {
     return response.status(400).end();
   }
 
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "Token invalid" });
+  }
+  const user = await User.findById(decodedToken.id);
+
   const blog = new Blog({
     ...body,
-    likes: body.likes ? body.likes : 0
+    likes: body.likes ? body.likes : 0,
+    author: user.id
   });
 
   try {
     const newBlog = await blog.save();
+    user.blogs = user.blogs.concat(newBlog.id);
+    await user.save();
     response.status(201).json(newBlog);
   } catch (error) {
     console.log("controller error: ", error);
@@ -28,11 +40,25 @@ blogsRouter.post("/", async (request, response, next) => {
 });
 
 blogsRouter.delete("/:id", async (request, response, next) => {
-  try {
-    await Blog.findByIdAndRemove(request.params.id);
-    response.status(204).end();
-  } catch (error) {
-    next(error);
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "Token invalid" });
+  }
+  const user = await User.findById(decodedToken.id);
+  if (!user) return response.status(401).json({ error: "Token invalid" });
+
+  const blog = await Blog.findById(request.params.id).populate("author");
+  if (blog.author.id.toString() === user.id.toString()) {
+    try {
+      await Blog.findByIdAndRemove(request.params.id);
+      user.blogs = user.blogs.filter(b => b.id !== blog.id);
+      await user.save();
+      response.status(204).end();
+    } catch (error) {
+      next(error);
+    }
+  } else {
+    return response.status(401).json({ error: "Token invalid" });
   }
 });
